Basic port checker working again
Port checker now returns port ids in numeric form now, accepting a path to save file and the name of the nation. It also checks to make sure there's a name-tag file to get those nation names for the user to select by and will run the generator if none is found. Some issues persist and returning names of ports still needs to be added.